refactor(infra): merge tirx_op/tirx_stmt into tile_primitive.h, rename Predicate to LambdaExpr
Consolidate the tile-primitive headers: merge tirx_stmt.h (TilePrimitiveCall)
and tirx_op.h (Op declarations + DispatchContext) into a single
tile_primitive.h, and move the reified-lambda object (formerly Predicate)
into the same header under the name LambdaExpr (type key tirx.LambdaExpr).

LambdaExpr keeps the existing semantics: vars carry SEqHashDefRecursive so
they act as lambda binders for structural eq/hash, and Apply substitutes
them with concrete indices (now via plain Substitute, since the select
predicate is int32-only and dtype legalization is a no-op). The Python
wrapper, FFI registration, builder, TRN codegen, printer dispatch, and
tests all follow the rename. Update all #include sites to the new header.
