+++ include/tvm/tirx/stmt_functor.h
@@ -31,7 +31,7 @@
 #include <tvm/tirx/expr_functor.h>
 #include <tvm/tirx/function.h>
 #include <tvm/tirx/stmt.h>
-#include <tvm/tirx/tirx_stmt.h>
+#include <tvm/tirx/tile_primitive.h>
 
 #include <unordered_map>
 #include <utility>
