+++ src/script/printer/doc_printer/python_doc_printer.cc
@@ -20,7 +20,7 @@
 #include <tvm/ffi/function.h>
 #include <tvm/ffi/reflection/registry.h>
 #include <tvm/script/printer/doc.h>
-#include <tvm/tirx/tirx_op.h>
+#include <tvm/tirx/tile_primitive.h>
 
 #include <algorithm>
 #include <cmath>
