+++ src/tirx/analysis/verify_tirx_well_formed.cc
@@ -30,7 +30,7 @@
 #include <tvm/tirx/op_attr_types.h>
 #include <tvm/tirx/stmt.h>
 #include <tvm/tirx/stmt_functor.h>
-#include <tvm/tirx/tirx_op.h>
+#include <tvm/tirx/tile_primitive.h>
 
 #include <exception>
 #include <optional>
